legacy/requirements.txt
curl_cffi==0.10.0
nicegui==2.14.1
python_dateutil==2.9.0.post0
